// common/enum.ts
export enum Gender {
    MALE = 'male',
    FEMALE = 'female',
    OTHER = 'other',
}

export enum FieldOfStudy {
    VeterinaryMedicine = 'Veterinary Medicine',
    AnimalScience = 'Animal Science',
    VeterinarySurgery = 'Veterinary Surgery',
    VeterinaryPathology = 'Veterinary Pathology',
    VeterinaryPharmacology = 'Veterinary Pharmacology',
    VeterinaryMicrobiology = 'Veterinary Microbiology',
    VeterinaryParasitology = 'Veterinary Parasitology',
    VeterinaryRadiology = 'Veterinary Radiology',
    VeterinaryAnatomy = 'Veterinary Anatomy',
    VeterinaryPhysiology = 'Veterinary Physiology',
    Other = 'Other',
}

export enum University {
    UniversityOfKarachi = 'University of Karachi',
    UniversityOfVeterinaryAndAnimalSciences = 'University of Veterinary and Animal Sciences',
    SindhAgricultureUniversity = 'Sindh Agriculture University',
    DowUniversityOfHealthSciences = 'Dow University of Health Sciences',
    LiaquatUniversityOfMedicalAndHealthSciences = 'Liaquat University of Medical and Health Sciences',
    JinnahSindhMedicalUniversity = 'Jinnah Sindh Medical University',
    AgaKhanUniversity = 'Aga Khan University',
    HamdardUniversity = 'Hamdard University',
    KarachiUniversityOfMedicalSciences = 'Karachi University of Medical Sciences',
    PakistanMaritimeUniversity = 'Pakistan Maritime University',
    InstituteOfBusinessAdministration = 'Institute of Business Administration',
    NEDUniversityOfEngineeringAndTechnology = 'NED University of Engineering and Technology',
    KarachiSchoolOfArt = 'Karachi School of Art',
    UniversityOfThePunjab = 'University of the Punjab',
    LahoreUniversityOfManagementSciences = 'Lahore University of Management Sciences',
    UniversityOfHealthSciencesLahore = 'University of Health Sciences, Lahore',
    BahauddinZakariyaUniversity = 'Bahauddin Zakariya University',
    GovernmentCollegeUniversityLahore = 'Government College University, Lahore',
    UniversityOfCentralPunjab = 'University of Central Punjab',
    UniversityOfAgricultureFaisalabad = 'University of Agriculture, Faisalabad',
    UniversityOfVeterinaryAndAnimalSciencesLahore = 'University of Veterinary and Animal Sciences, Lahore',
    Quaid_i_AzamUniversity = 'Quaid-i-Azam University',
    PirMehrAliShahAridAgricultureUniversity = 'Pir Mehr Ali Shah Arid Agriculture University',
    UniversityOfHealthSciences = 'University of Health Sciences',
    GovernmentCollegeUniversityFaisalabad = 'Government College University, Faisalabad',
    Other = 'Other',
}

export enum AnimalType {
    CAT = 'cat',
    DOG = 'dog',
}

export enum Status {
    PENDING = 'pending',
    CANCELLED = 'cancelled',
    DONE = 'done',
}

export enum PaymentStatus {
    PAID = 'paid',
    UNPAID = 'unpaid',
}

export enum Type {
    VIRTUAL = 'virtual',
    PHYSICAL = 'physical',
}

export enum UserType {
    VET = 'vet',
    USER = 'user',
}